-klasör yapılandırması ve Group Detail Page sayfası düzenlendi artık daha sade bir grup detay sayfası var -Snack bar değiştirildi. kendi tasarımımız mimari bozulmadan yapıldı
