+++ lib/features/groups/presentation/widgets/role_pill.dart
@@ -1,0 +1,26 @@
+import 'package:flutter/material.dart';
+
+class RolePill extends StatelessWidget {
+  const RolePill({super.key, required this.label, required this.color});
+  final String label;
+  final Color color;
+
+  @override
+  Widget build(BuildContext context) {
+    if (label.isEmpty) return const SizedBox.shrink();
+    final bg = color.withOpacity(.12);
+    final fg = color;
+    return Container(
+      padding: const EdgeInsets.symmetric(horizontal: 8, vertical: 4),
+      decoration: BoxDecoration(color: bg, borderRadius: BorderRadius.circular(999)),
+      child: Text(
+        label,
+        style: Theme.of(context).textTheme.labelSmall?.copyWith(
+          color: fg,
+          fontWeight: FontWeight.w600,
+          letterSpacing: .2,
+        ),
+      ),
+    );
+  }
+}
